requirements.txt
streamlit
pandas
python-docx
matplotlib
openpyxl
